Issue refs #1598: cancel, abort, progress tracking, session management implemented

+++ AGO.Reporting.Common/IProgressTracked.cs
@@ -4,7 +4,7 @@ namespace AGO.Reporting.Common
 {
     public interface IProgressTracker
     {
-        int PercentCompleted { get; }
+        byte PercentCompleted { get; }
 
         event EventHandler ProgressChanged;
     }
